refactor: Move presentation code to separate module
This commit moves the presentation layer code, including `MainActivity`, `SettingsViewModel`, `SettingsEvent`, and `SettingsState`, into a new `modules/presentation` package.

This restructuring improves the modularity of the codebase and separates concerns. The old files in `frontend` have been removed, and the `AndroidManifest.xml` has been updated to reflect the new package structure for `MainActivity`.

Additionally, the `AmoledDarkColors` theme is removed from `Theme.kt` as it's not currently used. The `SetAmoledTheme` event and the `amoledTheme` state in `SettingsState` are also removed. The `SettingsConstants` object is also slightly cleaned up.